refactor: add imageUtils
